Factor out JSON display code and improve readme
